requirements.txt
boto3==1.17.84
botocore==1.20.84
cryptography==3.4.7
fastapi==0.70.0
gunicorn==20.1.0
httptools==0.2.0
mangum==0.12.3
passlib==1.7.4
pydantic==1.8.2 --no-binary=pydantic
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
s3transfer==0.4.2
uvicorn==0.13.4
uvloop==0.15.2
pynamodb==5.1.0